+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v0.1.4: TBD
 ===================
 
 * Switched to using `jackson-datatype-guava` for JSON serialization/deserialization of Guava types.
+* Use `InstrumentedQueuedThreadPool` from `metrics-jetty`.
 
 
 v0.1.3: Jan 19 2012